Исправления по замечаниям к ДР_3
