+++ TinkoffFintech/Conversation/ViewModel/MessageCellModel.swift
@@ -2,7 +2,7 @@
 //  MessageCellModel.swift
 //  TinkoffFintech
 //
-//  Created by Anya on 28.09.2020.
+//  Created by Anya on 06.10.2020.
 //  Copyright © 2020 Anna Vondrukhova. All rights reserved.
 //
 
@@ -10,13 +10,4 @@ import Foundation
 
 struct MessageCellModel {
     let text: String
-}
-
-struct MessageViewModelFactory: ViewModelFactory {
-    typealias Model = Message
-    typealias ViewModel = MessageCellModel
-    
-    static func createViewModel(with model: Message) -> MessageCellModel {
-        return MessageCellModel(text: model.text)
-    }
 }
